requirements.txt
nicegui>=2.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
cryptography>=42.0.0
